projects/vision_agents/requirements.txt
langchain
langgraph
langsmith
langchain-core
langchain-openai
dotenv
rich
pymupdf
arxiv
rapidfuzz
markdownify